+++ app/src/main/java/com/android/rockages/kordx/services/i18n/Translation.g.kt
@@ -260,7 +260,6 @@ open class _Translation(private val container: _Container) {
         val Loseless: String,
         val UseMetaphonyMetadataDecoder: String,
         val PlayStore: String,
-        val GaplessPlayback: String,
         val AutoResumePlayback: String,
         val GridColumns: String,
         val CaseSensitiveSorting: String,
@@ -496,7 +495,6 @@ open class _Translation(private val container: _Container) {
     val Loseless: String get() = container.keys.Loseless
     val UseMetaphonyMetadataDecoder: String get() = container.keys.UseMetaphonyMetadataDecoder
     val PlayStore: String get() = container.keys.PlayStore
-    val GaplessPlayback: String get() = container.keys.GaplessPlayback
     val AutoResumePlayback: String get() = container.keys.AutoResumePlayback
     val GridColumns: String get() = container.keys.GridColumns
     val CaseSensitiveSorting: String get() = container.keys.CaseSensitiveSorting